+++ tests/test_checker/test_noqa.py
@@ -288,6 +288,7 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS612': 1,
     'WPS613': 1,
     'WPS614': 1,
+    'WPS615': 2,
 })
 
 # Violations which may be tweaked by `i_control_code` option: